Refactor connection.py to use urllib3
Refactor connection.py to use urllib3 due to the limitations of httplib2. We were unable to support self-signed certificates even if they were considered trusted. This refactor allowed me to move common functionality in a single class that can be used to make request. This greatly reduced the boilerplate code in the codebase.