Als replace http client library with something that supports unsigned certs (#10)
* Add parameter to set check_hostname according to !AllowSelfSigned

* Refactor connection.py to use urllib3

Refactor connection.py to use urllib3 due to the limitations of httplib2. We were unable to support self-signed certificates even if they were considered trusted. This refactor allowed me to move common functionality in a single class that can be used to make request. This greatly reduced the boilerplate code in the codebase.

* Refactor unit test to use urllib3

All unit test needed to be updated to use the urllib3 library. I took this as an opportunity to refactor duplicate code and functionality. This includes creating a function that captures stdout and moving common strs up to the setUp method.

* Revert version change

* Update _request to handle MaxRetryError

During testing, I found that a bad SSL certificate will throw a MaxRetryError. Additionally, I consolidated all the exception message to use the same exception block.

* Refactor move result init in handleResponse

* Update profile to correct handle errors

There was a hasattr missing after the first request. We only want to continue if the first request was successful.